fix: modularize types
