+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "wtw-accessible-react-datepicker",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "description": "An accessible datepicker with all the keyboard bindings, compatible with React",
   "keywords": [
     "datepicker",
@@ -25,7 +25,9 @@
   },
   "peerDependencies": {
     "react": "^17.0.2",
-    "wtw-icons": "^1.13.0"
+    "wtw-icons": "^1.13.0",
+    "focus-trap-react": "^8.5.0",
+    "styled-components": "^5.2.3"
   },
   "license": "MIT",
   "devDependencies": {
@@ -53,10 +55,6 @@
     "ts-jest": "^26.5.5",
     "typescript": "^4.2.3",
     "wtw-icons": "^1.14.0"
-  },
-  "dependencies": {
-    "focus-trap-react": "^8.5.0",
-    "styled-components": "^5.2.3"
   },
   "engines": {
     "node": "14",
