+++ src/DatePicker.tsx
@@ -9,7 +9,9 @@ import styled from "styled-components";
 import Calendar from "./Calendar";
 import { months as defaultMonths, days as defaultDays } from "./utils/defaults";
 import Header from "./Header";
-import { SelectedDates, StyleConfig, Tuple } from "./types";
+import { StyleConfig } from "./types/StyleConfig";
+import { SelectedDates } from "./types/SelectedDates";
+import { Tuple } from "./types/Tuple";
 import { useDateSelector } from "./utils/hooks/useDateSelector";
 import FocusTrap from "focus-trap-react";
 import { useOnClickOutside } from "./utils/hooks/useOnClickOutside";
