+++ src/index.ts
@@ -1,6 +1,8 @@
 export { default as Datepicker, DatePickerProps } from "./DatePicker";
 export { useToggle } from "./utils/hooks/useToggle";
-export { SelectedDates, Tuple } from "./types";
 export { generateDay } from "./utils/funcs/generateDay";
 export { generateDate } from "./utils/funcs/generateDate";
 export { differenceBetweenDays } from "./utils/funcs/differenceBetweenDays";
+export type { SelectedDates } from "./types/SelectedDates";
+export type { Day } from "./types/Day";
+export type { Tuple } from "./types/Tuple";
