+++ src/types/Day.ts
@@ -1,0 +1,5 @@
+export interface Day {
+  month: number;
+  day: number;
+  year: number;
+}
