+++ src/types/SelectedDates.ts
@@ -1,0 +1,3 @@
+import { useDateSelector } from "../utils/hooks/useDateSelector";
+
+export type SelectedDates = ReturnType<typeof useDateSelector>["selected"];
