+++ src/types/Tuple.ts
@@ -1,0 +1,3 @@
+export type Tuple<TItem, TLength extends number> = [TItem, ...TItem[]] & {
+  length: TLength;
+};
