+++ src/utils/consts.ts
@@ -1,4 +1,4 @@
-import { Tuple } from "../types";
+import { Tuple } from "../types/Tuple";
 import { isLeapYear } from "./funcs/isLeapYear";
 
 export const maxDate = (year: number): Tuple<number, 12> => [
