+++ src/utils/funcs/generateButtonId.ts
@@ -1,4 +1,4 @@
-import { Day } from "../../types";
+import { Day } from "../../types/Day";
 
 export const generateButtonId = (day: Day) =>
   `${day.day}-${day.month}-${day.year}`;
